Move `mock_lt_status`
